requirements.txt
semantic-kernel
pydantic
pandas
azure-identity
python-dotenv
azure-search-documents
